promene migracija,modela,i kontrolera(Ispravljanje gresaka)

+++ domaci1/database/migrations/2024_12_23_114101_add_role_id_to_users_table.php
@@ -13,9 +13,12 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->foreignId('role_id')->nullable()->constrained('roles')->onDelete('set null');
-        });
+        if (!Schema::hasColumn('users', 'role_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                // Dodavanje role_id kolone
+                $table->foreignId('role_id')->nullable()->constrained('roles')->onDelete('set null');
+            });
+        }
     }
 
     /**
@@ -26,7 +29,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            //
+            // Brisanje strane veze i kolone role_id
+            $table->dropForeign(['role_id']);
+            $table->dropColumn('role_id');
         });
     }
 };
